Migrate index test actions to actions package
