+++ tests/action/get_indexes.go
@@ -1,0 +1,138 @@
+// Copyright 2025 Democratized Data Foundation
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package action
+
+import (
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/sourcenetwork/immutable"
+
+	"github.com/sourcenetwork/defradb/client"
+	"github.com/sourcenetwork/defradb/tests/state"
+)
+
+// GetIndexes will attempt to get the indexes from the given collection
+// using the collection api.
+type GetIndexes struct {
+	stateful
+
+	// NodeID may hold the ID (index) of a node to get the indexes from.
+	//
+	// If a value is not provided the indexes will be retrieved from the first node.
+	NodeID immutable.Option[int]
+
+	// The identity of this request. Optional.
+	//
+	// If node acp is enabled, identity will be used to check if this operation can be performed.
+	Identity immutable.Option[state.Identity]
+
+	// The collection for which the indexes should be retrieved.
+	CollectionID int
+
+	// The expected indexes to be returned.
+	ExpectedIndexes []client.IndexDescription
+
+	// Any error expected from the action. Optional.
+	//
+	// String can be a partial, and the test will pass if an error is returned that
+	// contains this string.
+	ExpectedError string
+}
+
+var _ Action = (*GetIndexes)(nil)
+var _ Stateful = (*GetIndexes)(nil)
+
+func (a *GetIndexes) Execute() {
+	if len(a.s.Nodes) == 0 {
+		return
+	}
+
+	var expectedErrorRaised bool
+
+	nodeIDs, _ := getNodesWithIDs(a.NodeID, a.s.Nodes)
+	for _, nodeID := range nodeIDs {
+		collection := a.s.Nodes[nodeID].Collections[a.CollectionID]
+
+		ctx := getContextWithIdentity(a.s.Ctx, a.s, a.Identity, nodeID)
+		actualIndexes, err := collection.GetIndexes(ctx)
+		resetStateContext(a.s)
+
+		if assertError(a.s.T, err, a.ExpectedError) {
+			expectedErrorRaised = true
+			continue
+		}
+
+		assertIndexesListsEqual(a.ExpectedIndexes, actualIndexes, a.s.T)
+	}
+
+	assertExpectedErrorRaised(a.s.T, a.ExpectedError, expectedErrorRaised)
+}
+
+func assertIndexesListsEqual(
+	expectedIndexes []client.IndexDescription,
+	actualIndexes []client.IndexDescription,
+	t require.TestingT,
+) {
+	toNames := func(indexes []client.IndexDescription) []string {
+		names := make([]string, len(indexes))
+		for i, index := range indexes {
+			names[i] = index.Name
+		}
+		return names
+	}
+
+	require.ElementsMatch(t, toNames(expectedIndexes), toNames(actualIndexes))
+
+	toMap := func(indexes []client.IndexDescription) map[string]client.IndexDescription {
+		resultMap := map[string]client.IndexDescription{}
+		for _, index := range indexes {
+			resultMap[index.Name] = index
+		}
+		return resultMap
+	}
+
+	expectedMap := toMap(expectedIndexes)
+	actualMap := toMap(actualIndexes)
+	for key := range expectedMap {
+		assertIndexesEqual(expectedMap[key], actualMap[key], t)
+	}
+}
+
+func assertIndexesEqual(expectedIndex, actualIndex client.IndexDescription, t require.TestingT) {
+	assert.Equal(t, expectedIndex.Name, actualIndex.Name, "index name mismatch")
+	assert.Equal(t, expectedIndex.ID, actualIndex.ID, "index id mismatch")
+
+	toNames := func(fields []client.IndexedFieldDescription) []string {
+		names := make([]string, len(fields))
+		for i, field := range fields {
+			names[i] = field.Name
+		}
+		return names
+	}
+
+	require.ElementsMatch(t, toNames(expectedIndex.Fields), toNames(actualIndex.Fields), "index fields' names mismatch")
+
+	toMap := func(fields []client.IndexedFieldDescription) map[string]client.IndexedFieldDescription {
+		resultMap := map[string]client.IndexedFieldDescription{}
+		for _, field := range fields {
+			resultMap[field.Name] = field
+		}
+		return resultMap
+	}
+
+	expectedFieldsMap := toMap(expectedIndex.Fields)
+	actualFieldsMap := toMap(actualIndex.Fields)
+	for fieldName := range expectedFieldsMap {
+		assert.Equal(t, expectedFieldsMap[fieldName].Descending, actualFieldsMap[fieldName].Descending,
+			"index field %s descending mismatch", fieldName)
+	}
+}
