+++ tests/integration/acp/nac/index_list_test.go
@@ -49,7 +49,7 @@ func TestNAC_GatesIndexList_AuthorizedIdentity_AllowAccess(t *testing.T) {
 			},
 
 			// This should work as the identity is authorized.
-			testUtils.GetIndexes{
+			&action.GetIndexes{
 				Identity:        testUtils.ClientIdentity(1),
 				CollectionID:    0,
 				ExpectedIndexes: []client.IndexDescription{},
@@ -81,7 +81,7 @@ func TestNAC_GatesIndexList_NoIdentity_NotAuthorizedError(t *testing.T) {
 			},
 
 			// We haven't authorized non-identities. So, this should error.
-			testUtils.GetIndexes{
+			&action.GetIndexes{
 				Identity:      testUtils.NoIdentity(),
 				CollectionID:  0,
 				ExpectedError: "not authorized to perform operation",
@@ -113,7 +113,7 @@ func TestNAC_GatesIndexList_WrongIdentity_NotAuthorizedError(t *testing.T) {
 			},
 
 			// Wrong user/identity will also not be authorized.
-			testUtils.GetIndexes{
+			&action.GetIndexes{
 				Identity:      testUtils.ClientIdentity(2),
 				CollectionID:  0,
 				ExpectedError: "not authorized to perform operation",
