+++ tests/integration/index/drop_test.go
@@ -37,7 +37,7 @@ func TestIndexDrop_ShouldNotHinderQuerying(t *testing.T) {
 						"age":	21
 					}`,
 			},
-			testUtils.DropIndex{
+			&action.DropIndex{
 				IndexName: "User_name_ASC",
 			},
 			testUtils.Request{
@@ -82,10 +82,10 @@ func TestIndexDrop_ShouldRemoveIndexFromCollection(t *testing.T) {
 						"age":	21
 					}`,
 			},
-			testUtils.DropIndex{
+			&action.DropIndex{
 				IndexName: "User_age_ASC",
 			},
-			testUtils.GetIndexes{
+			&action.GetIndexes{
 				CollectionID: 0,
 				ExpectedIndexes: []client.IndexDescription{
 					{
@@ -97,10 +97,10 @@ func TestIndexDrop_ShouldRemoveIndexFromCollection(t *testing.T) {
 					},
 				},
 			},
-			testUtils.DropIndex{
+			&action.DropIndex{
 				IndexName: "User_name_ASC",
 			},
-			testUtils.GetIndexes{
+			&action.GetIndexes{
 				CollectionID:    0,
 				ExpectedIndexes: []client.IndexDescription{},
 			},
@@ -129,7 +129,7 @@ func TestIndexDrop_IfIndexDoesNotExist_ReturnError(t *testing.T) {
 						"age":	21
 					}`,
 			},
-			testUtils.DropIndex{
+			&action.DropIndex{
 				CollectionID:  0,
 				IndexName:     "non_existing_index",
 				ExpectedError: "index with name doesn't exists. Name: non_existing_index",
