+++ tests/integration/utils.go
@@ -437,15 +437,6 @@ func performAction(
 	case UpdateWithFilter:
 		updateWithFilter(s, action)
 
-	case CreateIndex:
-		createIndex(s, action)
-
-	case DropIndex:
-		dropIndex(s, action)
-
-	case GetIndexes:
-		getIndexes(s, action)
-
 	case CreateEncryptedIndex:
 		createEncryptedIndex(s, action)
 
@@ -1063,100 +1054,6 @@ func refreshDocuments(
 	}
 }
 
-func getIndexes(
-	s *state.State,
-	action GetIndexes,
-) {
-	if len(s.Nodes) == 0 {
-		return
-	}
-
-	var expectedErrorRaised bool
-
-	nodeIDs, _ := getNodesWithIDs(action.NodeID, s.Nodes)
-	for _, nodeID := range nodeIDs {
-		collections := s.Nodes[nodeID].Collections
-		ctx := getContextWithIdentity(s.Ctx, s, action.Identity, nodeID)
-		err := withRetryOnNode(
-			s.Nodes[nodeID],
-			func() error {
-				actualIndexes, err := collections[action.CollectionID].GetIndexes(ctx)
-				if err != nil {
-					return err
-				}
-
-				assertIndexesListsEqual(action.ExpectedIndexes,
-					actualIndexes, s.T)
-
-				return nil
-			},
-		)
-		expectedErrorRaised = expectedErrorRaised ||
-			AssertError(s.T, err, action.ExpectedError)
-	}
-
-	assertExpectedErrorRaised(s.T, action.ExpectedError, expectedErrorRaised)
-}
-
-func assertIndexesListsEqual(
-	expectedIndexes []client.IndexDescription,
-	actualIndexes []client.IndexDescription,
-	t testing.TB,
-) {
-	toNames := func(indexes []client.IndexDescription) []string {
-		names := make([]string, len(indexes))
-		for i, index := range indexes {
-			names[i] = index.Name
-		}
-		return names
-	}
-
-	require.ElementsMatch(t, toNames(expectedIndexes), toNames(actualIndexes))
-
-	toMap := func(indexes []client.IndexDescription) map[string]client.IndexDescription {
-		resultMap := map[string]client.IndexDescription{}
-		for _, index := range indexes {
-			resultMap[index.Name] = index
-		}
-		return resultMap
-	}
-
-	expectedMap := toMap(expectedIndexes)
-	actualMap := toMap(actualIndexes)
-	for key := range expectedMap {
-		assertIndexesEqual(expectedMap[key], actualMap[key], t)
-	}
-}
-
-func assertIndexesEqual(expectedIndex, actualIndex client.IndexDescription, t testing.TB) {
-	assert.Equal(t, expectedIndex.Name, actualIndex.Name, "index name mismatch")
-	assert.Equal(t, expectedIndex.ID, actualIndex.ID, "index id mismatch")
-
-	toNames := func(fields []client.IndexedFieldDescription) []string {
-		names := make([]string, len(fields))
-		for i, field := range fields {
-			names[i] = field.Name
-		}
-		return names
-	}
-
-	require.ElementsMatch(t, toNames(expectedIndex.Fields), toNames(actualIndex.Fields), "index fields' names mismatch")
-
-	toMap := func(fields []client.IndexedFieldDescription) map[string]client.IndexedFieldDescription {
-		resultMap := map[string]client.IndexedFieldDescription{}
-		for _, field := range fields {
-			resultMap[field.Name] = field
-		}
-		return resultMap
-	}
-
-	expectedMap := toMap(expectedIndex.Fields)
-	actualMap := toMap(actualIndex.Fields)
-	for key := range expectedMap {
-		assert.Equal(t, expectedMap[key], actualMap[key], "index fields' values mismatch")
-	}
-}
-
 func patchCollection(
 	s *state.State,
 	action PatchCollection,
@@ -1707,75 +1604,6 @@ func updateWithFilter(s *state.State, action UpdateWithFilter) {
 			immutable.None[state.Identity](),
 		)
 	}
-}
-
-// createIndex creates a secondary index using the collection api.
-func createIndex(
-	s *state.State,
-	action CreateIndex,
-) {
-	nodeIDs, nodes := getNodesWithIDs(action.NodeID, s.Nodes)
-	for index, node := range nodes {
-		nodeID := nodeIDs[index]
-		collection := s.Nodes[nodeID].Collections[action.CollectionID]
-		indexDesc := client.IndexCreateRequest{
-			Name: action.IndexName,
-		}
-		if action.FieldName != "" {
-			indexDesc.Fields = []client.IndexedFieldDescription{
-				{
-					Name: action.FieldName,
-				},
-			}
-		} else if len(action.Fields) > 0 {
-			for i := range action.Fields {
-				indexDesc.Fields = append(indexDesc.Fields, client.IndexedFieldDescription{
-					Name:       action.Fields[i].Name,
-					Descending: action.Fields[i].Descending,
-				})
-			}
-		}
-
-		indexDesc.Unique = action.Unique
-		ctx := getContextWithIdentity(s.Ctx, s, action.Identity, nodeID)
-		err := withRetryOnNode(
-			node,
-			func() error {
-				_, err := collection.CreateIndex(ctx, indexDesc)
-				return err
-			},
-		)
-		if AssertError(s.T, err, action.ExpectedError) {
-			return
-		}
-	}
-
-	assertExpectedErrorRaised(s.T, action.ExpectedError, false)
-}
-
-// dropIndex drops the secondary index using the collection api.
-func dropIndex(
-	s *state.State,
-	action DropIndex,
-) {
-	var expectedErrorRaised bool
-
-	nodeIDs, nodes := getNodesWithIDs(action.NodeID, s.Nodes)
-	for index, node := range nodes {
-		nodeID := nodeIDs[index]
-		collection := s.Nodes[nodeID].Collections[action.CollectionID]
-
-		ctx := getContextWithIdentity(s.Ctx, s, action.Identity, nodeID)
-		err := withRetryOnNode(
-			node,
-			func() error {
-				return collection.DropIndex(ctx, action.IndexName)
-			},
-		)
-		expectedErrorRaised = AssertError(s.T, err, action.ExpectedError)
-	}
-
-	assertExpectedErrorRaised(s.T, action.ExpectedError, expectedErrorRaised)
 }
 
 func createEncryptedIndex(
